代码修改

+++ src/main/java/com/cdkj/loan/ao/impl/RepointDetailAOImpl.java
@@ -83,8 +83,8 @@ public class RepointDetailAOImpl implements IRepointDetailAO {
     }
 
     @Override
-    public Paginable<RepointDetail> queryRepointDetailPage(int start,
-            int limit, RepointDetail condition) {
+    public Paginable<RepointDetail> queryRepointDetailPage(int start, int limit,
+            RepointDetail condition) {
         return repointDetailBO.getPaginable(start, limit, condition);
     }
 
@@ -157,8 +157,8 @@ public class RepointDetailAOImpl implements IRepointDetailAO {
         Long fxAmount = getLong(data.getFxAmount());
         Long otherFee = getLong(data.getOtherFee());
         Long sxFee = 0L;// 收客户手续费合计：履约保证金+担保风险金+杂费
-        if (EServiceChargeWay.REPOINT.getCode().equals(
-            data.getServiceChargeWay())) {
+        if (EServiceChargeWay.REPOINT.getCode()
+            .equals(data.getServiceChargeWay())) {
             sxFee = lyAmount + fxAmount + otherFee;
         }
         Long gpsFee = 0L;// GPS费
@@ -166,15 +166,15 @@ public class RepointDetailAOImpl implements IRepointDetailAO {
             gpsFee = getLong(data.getGpsFee());
         }
         actualRepointAmount = repointAmount - sxFee - gpsFee;
-        if (ERepointType.SINGLE.getCode().equals(
-            carDealerProtocol.getReturnPointType())) {
+        if (ERepointType.SINGLE.getCode()
+            .equals(carDealerProtocol.getReturnPointType())) {
             actualRepointAmount = actualRepointAmount
                     - getLong(carDealerProtocol.getReturnPointFee());
         }
-        if (ERepointType.PERCENT.getCode().equals(
-            carDealerProtocol.getReturnPointType())) {
+        if (ERepointType.PERCENT.getCode()
+            .equals(carDealerProtocol.getReturnPointType())) {
             actualRepointAmount = AmountUtil.mul(actualRepointAmount,
-                (1 - getLong(carDealerProtocol.getReturnPointRate())));
+                (1 - getDouble(carDealerProtocol.getReturnPointRate())));
         }
         // 获得汽车经销商返点账号集合
         CollectBankcard condition = new CollectBankcard();
@@ -185,11 +185,10 @@ public class RepointDetailAOImpl implements IRepointDetailAO {
         for (CollectBankcard collectBankcard : list) {
             if (collectBankcard.getBankCode().equals(bank.getBankCode())) {
                 Double pointRate = getDouble(collectBankcard.getPointRate());
-                repointDetail
-                    .setUseMoneyPurpose(EUseMoneyPurpose.PROTOCOL_INNER
-                        .getCode());
-                repointDetail.setRepointAmount(AmountUtil.mul(
-                    actualRepointAmount, pointRate));// 实际返点金额*返点比例=实际给汽车经销商返点的金额
+                repointDetail.setUseMoneyPurpose(
+                    EUseMoneyPurpose.PROTOCOL_INNER.getCode());
+                repointDetail.setRepointAmount(
+                    AmountUtil.mul(actualRepointAmount, pointRate));// 实际返点金额*返点比例=实际给汽车经销商返点的金额
                 repointDetail.setAccountCode(collectBankcard.getCode());
                 repointDetail.setBenchmarkRate(benchmarkRate);
             }
